python_backend/requirements.txt
# ── Phase 1: Core scaffolding ──
fastapi
uvicorn
python-multipart

# ── Phase 2: Nutrition Engine ──
chromadb
sentence-transformers
rank_bm25
pandas

# ── Phase 3: Vision Engine ──
transformers
pillow

# ── Testing ──
pytest
httpx

# ── Phase 5: Multi-modal Endpoints ──
google-generativeai
openai-whisper
python-dotenv
